Move packing list entry resolvers into dedicated modules
